Update policy server state, disable policy rules
