+++ pkg/kubewarden/chart/kubewarden/admission/Rules.vue
@@ -5,6 +5,8 @@ import { removeAt } from '@shell/utils/array';
 
 import Loading from '@shell/components/Loading';
 
+import { KUBEWARDEN_APPS } from '../../../types';
+import { ARTIFACTHUB_PKG_ANNOTATION } from '../../../plugins/kubewarden-class';
 import Rule from './Rule';
 
 export default {
@@ -45,6 +47,12 @@ export default {
       return this.$store.getters[`${ this.currentProduct.inStore }/all`]('apigroup');
     },
 
+    disabledRules() {
+      const annotations = this.value.policy?.metadata?.annotations;
+
+      return !!annotations?.[ARTIFACTHUB_PKG_ANNOTATION] || annotations?.['meta.helm.sh/release-name'] === KUBEWARDEN_APPS.RANCHER_DEFAULTS;
+    },
+
     isView() {
       return this.mode === _VIEW;
     }
@@ -69,10 +77,11 @@ export default {
       <Rule
         ref="lastRule"
         v-model="rules[index]"
+        :disabled="disabledRules"
         :mode="mode"
         :api-groups="apiGroups"
       >
-        <template v-if="!isView" #removeRule>
+        <template v-if="!isView && !disabledRules" #removeRule>
           <button type="button" class="btn role-link p-0" @click="removeRule(index)">
             {{ t('kubewarden.policyConfig.rules.remove') }}
           </button>
@@ -80,7 +89,7 @@ export default {
       </Rule>
     </div>
 
-    <button v-if="!isView" type="button" class="btn role-tertiary add" @click="addRule">
+    <button v-if="!isView && !disabledRules" type="button" class="btn role-tertiary add" @click="addRule">
       {{ t('kubewarden.policyConfig.rules.add') }}
     </button>
   </div>
